Move Portfolio property from HomePage to App
