+++ src/pages/App/App.js
@@ -11,21 +11,39 @@ import ProjectDetailPage from '../ProjectDetailPage/ProjectDetailPage';
 // import AuthPage from '../AuthPage/AuthPage';
 import NavBar from '../../components/NavBar/NavBar';
 import Profile from '../../components/Profile/Profile';
+import LoginForm from '../../components/LoginForm/LoginForm';
+// import Portfolio from '../../components/Portfolio/Portfolio'
 
 
 // Helpers
 import { getUser } from '../../utilities/services/users';
-import LoginForm from '../../components/LoginForm/LoginForm';
+import { getPortfolioItems } from '../../utilities/api/portfolioItem'
+import { removePortfolioItem } from '../../utilities/api/portfolioItem';
 
 
 // import '../../assets/css/style.css';
 
 export default function App() {
   const [user, setUser] = useState(getUser());
+  const [portfolioItems, setPortfolioItems] = useState([])
+
+
+  useEffect (function () {
+    async function getPortfolioItemList() {
+        const portfolioItemList = await getPortfolioItems()
+        setPortfolioItems(portfolioItemList)
+    }
+    getPortfolioItemList()
+    },[])
+
+  async function deletePortfolioItem(item) {
+    // console.log(item)
+    const newPortfolioItemList = await removePortfolioItem(item)
+    setPortfolioItems(newPortfolioItemList)
+  }
 
   return (
     <>
-      {/* <Scripts/> */}
       {/* Mobile nav toggle button */}
       <i className="bi bi-list mobile-nav-toggle d-xl-none"></i>
 
@@ -38,8 +56,8 @@ export default function App() {
       
       <Routes>
           {/* Route components in here */}
-          <Route path="/" element={<HomePage user={user} />} />
-          <Route path="/projects/:projectName" element={<ProjectDetailPage user={user} />} />
+          <Route path="/" element={<HomePage user={user} portfolioItems={portfolioItems} setPortfolioItems={setPortfolioItems} deletePortfolioItem={deletePortfolioItem} />} />
+          <Route path="/projects/:projectName" element={<ProjectDetailPage user={user} portfolioItems={portfolioItems} />} />
           <Route path="/AdminLogin" element={<LoginForm user={user} setUser={setUser} />} />
       </Routes>
     </>
