+++ src/pages/Homepage/HomePage.jsx
@@ -14,16 +14,16 @@ import { getAbout } from '../../utilities/api/about'
 import { getHome } from '../../utilities/api/home'
 import { getSkills } from '../../utilities/api/skill'
 import { removeSkill } from '../../utilities/api/skill'
-import { getPortfolioItems } from '../../utilities/api/portfolioItem'
+// import { getPortfolioItems } from '../../utilities/api/portfolioItem'
 import { removePortfolioItem } from '../../utilities/api/portfolioItem'
 
 
-export default function HomePage({user}) {
+export default function HomePage({user, portfolioItems, setPortfolioItems, deletePortfolioItem}) {
     const [contact, setContact] =useState([])
     const [about, setAbout] = useState([])
     const [home, setHome] = useState([])
     const [skills, setSkills] = useState([])
-    const [portfolioItems, setPortfolioItems] = useState([])
+    // const [portfolioItems, setPortfolioItems] = useState([])
     useEffect (function () {
         async function getContacts() {
             const contacts = await getContact();
@@ -41,15 +41,15 @@ export default function HomePage({user}) {
             const skillList = await getSkills()
             setSkills(skillList)
         }
-        async function getPortfolioItemList() {
-            const portfolioItemList = await getPortfolioItems()
-            setPortfolioItems(portfolioItemList)
-        }
+        // async function getPortfolioItemList() {
+        //     const portfolioItemList = await getPortfolioItems()
+        //     setPortfolioItems(portfolioItemList)
+        // }
         getHomeInfo()
         getContacts()
         getAbouts()
         getSkillList()
-        getPortfolioItemList()
+        // getPortfolioItemList()
     },[])
     
     async function deleteSkill(skill) {
@@ -58,11 +58,11 @@ export default function HomePage({user}) {
         setSkills(newSkills)
     }
 
-    async function deletePortfolioItem(item) {
-        // console.log(item)
-        const newPortfolioItemList = await removePortfolioItem(item)
-        setPortfolioItems(newPortfolioItemList)
-    }
+    // async function deletePortfolioItem(item) {
+    //     // console.log(item)
+    //     const newPortfolioItemList = await removePortfolioItem(item)
+    //     setPortfolioItems(newPortfolioItemList)
+    // }
 
     return (
         <>
@@ -71,7 +71,7 @@ export default function HomePage({user}) {
                 <About user={user} about={about} setAbout={setAbout}/>
                 <SkillsList user={user} skills={skills} setSkills={setSkills} deleteSkill={deleteSkill} />
                 <Resume user={user}/>
-                <Portfolio user={user} portfolioItems={portfolioItems} setPortfolioItems={setPortfolioItems} deletePortfolioItem={deletePortfolioItem} />
+                <Portfolio user={user} portfolioItems={portfolioItems} setPortfolioItems={setPortfolioItems}  deletePortfolioItem={deletePortfolioItem}/>
                 <Contact user={user} contact={contact} setContact={setContact}/>
             </main>
         </>
